laatste wijzigingen

+++ src/main/java/nl/kevinvanschaijk/rest/Playlist/PlaylistController.java
@@ -1,4 +1,4 @@
-package nl.kevinvanschaijk.rest.Playlist;
+package nl.kevinvanschaijk.rest.playlist;
 
 import nl.kevinvanschaijk.service.entity.playlist.Playlist;
 import nl.kevinvanschaijk.service.entity.track.Track;
@@ -9,6 +9,8 @@ import javax.naming.AuthenticationException;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+
+import static org.apache.cxf.common.util.StringUtils.isEmpty;
 
 
 @Path("/playlists")
@@ -21,10 +23,14 @@ public class PlaylistController {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAllPlaylists(@QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.getPlaylists(token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.ok().entity(playlistService.getPlaylists(token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
@@ -34,10 +40,14 @@ public class PlaylistController {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAllTracksInPlaylist(@PathParam("id") int id, @QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.getPlaylistTracks(id, token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.ok().entity(playlistService.getPlaylistTracks(id, token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
@@ -47,10 +57,14 @@ public class PlaylistController {
     @Produces(MediaType.APPLICATION_JSON)
     public Response addPlaylist(Playlist playlist, @QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.addPlaylist(playlist, token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.status(Response.Status.CREATED).entity(playlistService.addPlaylist(playlist, token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
@@ -58,12 +72,16 @@ public class PlaylistController {
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{id}")
-    public Response deletePlaylist(@PathParam("id") String id, @QueryParam("token") String token) {
+    public Response deletePlaylist(@PathParam("id") int id, @QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.deletePlaylist(id, token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.ok().entity(playlistService.deletePlaylist(id, token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
@@ -71,12 +89,16 @@ public class PlaylistController {
     @PUT
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{id}")
-    public Response updatePlaylist(Playlist playlist, @PathParam("id") String id, @QueryParam("token") String token) {
+    public Response updatePlaylist(Playlist playlist, @PathParam("id") int id, @QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.updatePlaylist(playlist, id, token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.ok().entity(playlistService.updatePlaylist(playlist, id, token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
@@ -86,10 +108,14 @@ public class PlaylistController {
     @Path("/{playlistId}/tracks/{trackId}")
     public Response deletePlaylistTrack(@PathParam("playlistId") int playlistId, @PathParam("trackId") int trackId, @QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.deletePlaylistTrack(playlistId, trackId, token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.ok().entity(playlistService.deletePlaylistTrack(playlistId, trackId, token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
@@ -100,10 +126,14 @@ public class PlaylistController {
     @Path("/{playlistId}/tracks")
     public Response addPlaylistTrack(Track track, @PathParam("playlistId") int playlistId, @QueryParam("token") String token) {
         Response response = null;
-        try {
-            response = Response.ok().entity(playlistService.addPlaylistTrack(track, playlistId, token)).build();
-        } catch (AuthenticationException e) {
-            response = Response.status(Response.Status.FORBIDDEN).build();
+        if (!isEmpty(token)) {
+            try {
+                response = Response.status(Response.Status.CREATED).entity(playlistService.addPlaylistTrack(track, playlistId, token)).build();
+            } catch (AuthenticationException e) {
+                response = Response.status(Response.Status.FORBIDDEN).build();
+            }
+        } else {
+            response = Response.status(Response.Status.BAD_REQUEST).build();
         }
         return response;
     }
